requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
lxml==4.6.3
paho-mqtt==1.5.1
python-libtorrent-bin==2.0.2
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
